+++ tests/call_fn.rs
@@ -146,7 +146,7 @@ fn test_fn_ptr_raw() -> Result<(), Box<EvalAltResult>> {
                 let value = args[2].clone();
                 let this_ptr = args.get_mut(0).unwrap();
 
-                fp.call_dynamic(&context, Some(this_ptr), [value])
+                fp.call_raw(&context, Some(this_ptr), [value])
             },
         );